#15060: host side code support for new all gather
#15060: moved new all gather to separate all gather v2 codebase

#15060: WIP code for adding line fabric to all gather v2

#15060: repro for edmLineFabric unhashable type

#15060: fixed error on passing fabric into all gather op

WIP: update all-gather async to work with fabric

teardown and syncmode support added (untested)

misc changes:
- add command processor header for common command processing functionality
- local chip only fabric build added

Missing: lifetime globals for:
- buffer addressing
- completion/sync mode semaphores

For the time being, both buffers and semaphore are allocated locally and
are assumed to line up on the same cores/addresses between chips because
we are in a single op test environment. THIS IS NOT ACCEPTABLE AS A
GENERAL SOLUTION and will need to be fixed before we merge. However,
it's enough to unblock progress on the baseline op functionality.

#0: wip for bring up 2 chip all gather