#14977: conv config to use higher cores. (#15962)
### Ticket
#14977

### Problem description
The convolution in these layers were using 8 cores only. Changed the
config to row major so that they can have use more cores.

This also improves performance as some of the tensor layout conversion
are not required anymore.
if the output is in bfloat16 and output_layout is ROW_MAJOR it can take
partial tile as shard shape and will use all 64 cores.