#15018: Increase linking of program binary multicasts
The BRISC, NCRISC, and set of all TRISCs each have their own entry in
kq_transfer_info.dst_base_addrs, which prevents the current code from linking
their transactions. Modify the code to avoid unlinking in the case where writes
to the same set of cores happen to be contiguous (which happens to be always).

This reduces the cost of linking and will help a lot more in the future, once
we put barriers between unlinked mcasts.