#15018: Re-enable async dispatch and workaround NOC hang
When async dispatch was previously enabled, we hit a hang in LLAMA that seems
to occur when doing the path reserve for an mcast and a previous mcast is still
active. To work around this, force write barriers before mcasts that aren't
linked to earlier mcasts. The dispatcher is the only core that sends mcasts
over VC 5, and even with multiple CQs only one can be mcasting programs at a
time.