#16547: Switch to google benchmark for pgm dispatch tests
The current output format works well with a smaller number of tests, but with a
large number of tests it's hard to connect the resulting number to the test
that created it. This particular hinders storing test results in a database and
comparing them across runs.

Switch to using the google benchmark framework, which outputs a JSON file
containing the results of all tests. By default, running the binary runs all
the benchmarks from sweep_pgm_dispatch.sh. The set of tests to be run can be
filtered using the `--benchmark_filter=<regex>` command-line argument.

One-off test cases can be run by passing `--custom` and the command line arguments as before.