Conv2dConfig reallocate_halo_output default to true
Idea behind this flag is to enable reallocation of
halo output buffer in place of input buffer if case
deallocate_activation flag is also set.
In this case halo output is moved up and memory
fragmentation is reduced.

In case deallocate_activation is not set reallocate_halo_output
won't have an effect as ttnn::move will be a no-op
since input buffer is not deallocated.

Current problem is that if user sets deallocate_activation
it doesn't help them with memory issue unless they
set reallocate_halo_output as well.

This change is addressing this by setting reallocate_halo_output
to true by default.

This helps by increassing pass rate in ttnn torch
traces.