#15605: Only force-stall ethernet programs on earlier ethernet programs
Keep track of when the last program using active ethernet cores was dispatched,
so we can wait on that program before sending out binaries. This is better than
always waiting on the immediate previous program, since in most cases we don't
run programs on the ethernet cores back-to-back.

This is a reland of the original patch that includes a fix for a bug in
WorkerConfigBuffer that caused hangs in
RandomProgramTraceFixture.TensixActiveEthTestSimpleProgramsTrace