#16171: Assert that NCRISC NOC is idle at kernel end.
Outstanding NOC transactions can cause incorrect values to be set when
noc_local_state_init is called before starting the next kernel. This hasn't
been a problem until now because dispatch took long enough that all outstanding
NOC transactions would complete by that point, but as we optimize the
dispatcher that will soon no longer be the case.

To fix that, in DM_DEDICATED_NOC mode we can assert at the end of the
kernel_launch that no transactions are outstanding. In DM_DYNAMIC_NOC mode
brisc.cc waits for transactions to copmlete, so we don't need to worry about
that mode. We could delay the wait until later, but brisc.cc doesn't have
access to the NCRISC's local NOC state when DM_DEDICATED_NOC is in use. We
could also have the ncrisck.cc wait for transactions instead of just asserting
that they're completed, but it could be faster for kernels to skip waits for
transactions types they don't use.

Some important NCRISC kernels have been fixed to ensure they perform a full barrier.

idle_erisck also has the checks added since it can have the same problem
(especially with slave eriscs).