Add support for tiled indices via padding/alignment aware embedding kernel (tiled indices only) (#16296)
### Ticket
[#14915](https://github.com/tenstorrent/tt-metal/issues/14915)

### Problem description
uint32 untilize not accurate, so padding aware kernel was needed as
opposed to standard conversion

### What's changed
A lot of changes to look at, first is the split_work_to_cores_aligned
function.

I made this function to help split work to cores such that the units per
core is aligned with some value. In the new embedding kernel, I align
with the face_height of a tile. After developing this kernel further, I
believe it's possible to use the standard split_work function for more
parallelism, but I figured I'd keep it for now as it could prove useful
for perf in the future where we want to make sure our split work is
aligned. (You can use the new aligned function as an immediate drop in
replacement for the same return values.)

As for the new program factory and kernel, the principle of weights
writing is identical to the RM kernel, and this means we want to read
the indices such that we are accessing the tensor in an RM fashion, you
could compare it simply to a parallelized print. A lot of the work is
simply offset algebra, and I tried to make sure to use const expr for
any tile or face heights, do let me know if there's anywhere I don't!

With this change, sweeps for embedding go from 46% -> 92% due to the
support of tiled indices.

### Checklist
- [x] [Post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/12641263680)
- [x] [Model regression CI testing
passes](https://github.com/tenstorrent/tt-metal/actions/runs/12589016962)