Resolve reduce-scatter-async sharded tensor correctness bug & hang (#16548)
Fix sharding support in reduce-scatter-async

the implementation was buggy because it was sharing kernel IDs for
partial reducer workers in all cases. However, for end-of-line devices,
the designated partial reducer cores have different input and output
tensor configurations, which means their address generators are not
sharable. Addrgens are compile time defined (and for sharded, the shard
spec info is also specified at compile time), so these kernels are
incompatible.

The end result was that some workers would have the incorrect address
generator.

The issue has been resolved and kernels are created correctly to account
for this.