Sever TT-Train's dependency on TT-Metalium's tests (#16685)
### Ticket
#16678

### Problem description
TT-Train is depending on a bit from TT-Metalium's tests. This was
exposed when cleaning up include directories.

### What's changed
Turns out that header was used to:
1) Populate an unused variable
2) Print a log message.

If that log message in the sample code is important, we can find another
way to get the info. This diff assumes that log message isn't crucial.

### Checklist
- [ ] Post commit CI passes
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [ ] New/Existing tests provide coverage for changes