Remove redundant bank offset from destination address in `ttnn.reshard`
Previously, the bank offset was added twice (once on host and once by
`get_noc_addr_from_bank_id`). We never noticed this issue because the
test was not actually asserting on the correctness check.

This change also fixes a Watcher assert in the kernel due to incorrect
barrier which would trigger when doing a DRAM->L1 reshard.

Adds a missing assert to reshard unit test.