Fix CB allocation warnings from ttnn.reshard
This change is motivated by convs and unet shallow model.
Conv2d calls reshard from within the conv2d composite op.
In unet model we are hitting many warnings for CB allocation in release,
and asserts in debug which make it hard to work on the model.

With this change all tests in test_reshard.py pass in debug model as well.