Instead of relying on parallelisaion strategy and pick sharding scheme which maximises tensix core count in order to provide more L1 space for the op, use L1 estimates to pick the sharding scheme, which minimises the L1 usage per tensix core.
This way auto-sharding is more robust and
doesn't include logic to override sharding with
max core count in some cases.
For this reason this strategy is a better fit
for BH.

Current L1 estimates only include conv2d device op
CB allocations and tensor output for conv2d device
op, meaning halo device op is ignored.

Main shortcomming of the current implementation is
that output buffer for halo op is ignored and that
one can be critical in some cases. This is a main
ToDo.

This change improves pass rate on pytorch and
tt-forge traces for both WH and BH.