Address some CMake warnings (#16993)
### Ticket
#14001

### Problem description
The CMake output is littered with warnings, which makes it difficult to
see actual errors.

### What's changed
Addressed some of the warnings.  Still more to do. To be continued...
Mostly addressed by telling CMake that we've tested the CMake files
against newer CMake version and it's okay to enable the policies
introduced since the minimum supported version.
In some cases this was done by grabbing a newer release of the library,
and in others we patch the `cmake_minimum_required` line ourselves as we
pull it in.

### Checklist
- [x] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/12918179996