PR Gate workflow as the nucleus to build up a pre-merge sanity check (#17097)
### Ticket
#17095

### Problem description
We don't even compile our repo automatically in PRs. Re require devs to
navigate the maze of GH to find the right button to mash. And we can't
just auto-run APC because that's crazy long (and heavy on the infra).

### What's changed
A new workflow that does a simple build. We'll expand later, but with an
eye on robustness and speed.