#17307: Run NCRISC FW from L1 on Wormhole
This saves IRAM space for the kernel, and allows later patches to optimize
NCRISC initialization more.

This is only enabled on Wormhole because a memory arbiter bug prevents NCRISC
from safely reading instructions from L1 on Grayskull.

We need to reset NCRISC when switching from executing in L1 to executing in
IRAM, because otehrwise we were seeing odd behavior where it was sometimes
misexecuting kernel code.