decouple channel slot pointers for edm fabric sender channel
The sender side channels are able to manage the progress of the following pointers independent from each other:

1) Sender side pointers to receiver buffer, these are shared between both sender channels because they are used to manage writes into the receiver buffer
- wrptr
- ackptr
- compl ptr

2) Per sender channel internal pointers. These are used by sender channel to understand internal capacity and progress through its own channel buffer.
* wrptr
  * advances when worker sender indicates new data is available (by incrementing the worker <-> edm interface wrptr)
  * trails worker -> EDM wrptr
  * must make sure not to overflow/trample/race-ahead of ackptr and rdptr
* ackptr
  * advances when receiving an ack signal from receiver
  * trails wrptr
* rdptr
  * advances when receiving a completion signal from receiver
  * trails wrptr

The primary benefit of this change is that all of the above can advance at their own rate, until backpressured. The previous implementation treated all of these the same and instead managed progress through the channel via a state machine. This prior implementation essentially forces progress of these three conceptually separate pointers to advance in lock-step, which is detrimental to performance, particularly for latency.