Setting venv specifically for Wormhole
The WH_ARCH_YAML is a variable used by tt-metal to instruct the runtime to dispatch specific kernels to specific tensix cores on the core grid. It is required for n300 and not n150 because n300 cards have different "harvesting" specs. Specifically they harvest two rows of the tensix grid instead of 1 row like the n150 does.
Setting this variable to that value informs tt-metal of the location of the specific ETH cores to use for dispatch.