Add support for parallelization along the width for untilize with unpadding  (#17538)
### Ticket
Link to Github Issue
https://github.com/tenstorrent/tt-metal/issues/17537

### Problem description
Currently, the untilize with unpadding implementation supports
parallelization only along the height dimension. This affects perf for
wide tensors, as they are mapped to a limited number of cores.

### What's changed
In this PR, we introduce support for parallelizing the untiling
operation along the width dimension, similar to tilize with padding. The
operation executes the parallelization over the dimension with the
larger number of tiles.
In future versions: 
- we want the operation to support the parallelization along both
dimensions simultaneously
- we want the compute kernel to support the processing of an entire
column block at once instead of one tile at a time

For the tests added in test_to_layout.py, the kernel duration of the
previous implementation is around 1.8 to 24.8 times larger than the
current implementation  

### Checklist
- [x] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/13121055787
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [ ] New/Existing tests provide coverage for changes