Update EDM Fabric flow control between sender and receiver channels over Ethernet to use stream registers and eth reg write (#17577)
### Problem description
Flow control between sender and receiver channels is currently done by
embedding the information within the channel buffer directly. As a
result several tradeoffs are introduced:
- Fixed max packet size
- Tradeoff decision to send larger packet with garbage data or send two
packets for flow control
- Numerous additional arithmetic calculations to determine offsets in
the channel buffer to find next `channel_syncs` (the embedded memory
containing flow control info per packet).

### What's changed
Make use of stream registers to track credit exchange between
sender/receiver in EDM fabric. These stream registers can be written two
and have an atomic increment feature on write. This includes all write
type (local and from eth_reg_write). This saves us lots of calculations
that were previously spent on computing offsets into channel buffers for
reading of channel syncs.

The flow control is implemented through the use of several stream
registers: one per conceptual pointer being tracked.
In total there are 5 such pointers (registers) in use:
1) to receiver channel packets sent
- Incremented by sender (via eth_reg_write) by the number of buffer
slots written. In practice, this means it is
    incremented once per packet
2) to sender 0 packets acked
- Incremented by receiver for every new packet from channel 0 that it
sees
3) to sender 1 packets acked
- Incremented by receiver for every new packet from channel 1 that it
sees
4) to sender 0 packets completed
- Incremented by receiver for every packet from channel 0 that it
completes processing for
5) to sender 1 packets completed
- Incremented by receiver for every packet from channel 1 that it
completes processing for

All updates to the remote copy of the pointer are performed through the
eth_reg_write API, which enables the caller to embed the payload in the
command submitted to the eth tx command queue.

### Note on Source Channel Identification
The receiver channel needs to know who sent each given packet so it can
correctly direct acks to the right producer sender channel. The new
implementation here is to require the sender channel to write their
channel ID to the new sender channel field in the packet header. This
field is only used by the sender and receiver channel and is otherwise
reserved. The receiver reads this part of the header to understand which
of the above ack/completion registers to `eth_reg_write` to.